Added upload image to repository
This web page features a registration form allowing users to input their personal details like name, birthdate, email, and password. Additionally, users can upload an image. Once the form is submitted, the provided information, including the image, is securely stored in a SQL Server database. The page is designed to be user-friendly and compatible with various devices.